@@ -1,10 +1,18 @@
 import setuptools
 
+
+def open_requirements(path):
+    with open(path) as f:
+        requires = [
+            r.split('/')[-1] if r.startswith('git+') else r
+            for r in f.read().splitlines()]
+    return requires
+
+
 # Get requirements from requirements.txt, stripping the version tags
-with open('requirements.txt') as f:
-    requires = [
-        r.split('/')[-1] if r.startswith('git+') else r
-        for r in f.read().splitlines()]
+requires = open_requirements('requirements.txt')
+tests_requires = open_requirements('extra_requirements/requirements-tests.txt')
+doc_requirements = open_requirements('extra_requirements/requirements-docs.txt')
 
 with open('README.md') as file:
     readme = file.read()
@@ -13,7 +21,7 @@ with open('HISTORY.md') as file:
     history = file.read()
 
 setuptools.setup(name='straxen',
-                 version='0.14.5',
+                 version='1.2.8',
                  description='Streaming analysis for XENON',
                  author='Straxen contributors, the XENON collaboration',
                  url='https://github.com/XENONnT/straxen',
@@ -21,29 +29,32 @@ setuptools.setup(name='straxen',
                  long_description_content_type="text/markdown",
                  setup_requires=['pytest-runner'],
                  install_requires=requires,
-                 tests_require=requires + [
-                     'tensorflow',
-                     'pytest',
-                     'hypothesis',
-                     'boltons'],
-                 python_requires=">=3.6",
+                 tests_require=requires + tests_requires,
+                 # Not testing py3.6 or py3.7 #616
+                 python_requires=">=3.8",
                  extras_require={
-                     'docs': ['sphinx',
-                              'sphinx_rtd_theme',
-                              'nbsphinx',
-                              'recommonmark',
-                              'graphviz'],
+                     'docs': doc_requirements,
                      'microstrax': ['hug'],
                  },
-                 scripts=['bin/bootstrax', 'bin/straxer', 'bin/fake_daq',
-                          'bin/microstrax', 'bin/ajax',
-                          'bin/refresh_raw_records'],
-                 packages=setuptools.find_packages(),
+                 scripts=[
+                     'bin/bootstrax',
+                     'bin/straxer',
+                     'bin/fake_daq',
+                     'bin/microstrax',
+                     'bin/ajax',
+                     'bin/refresh_raw_records',
+                 ],
+                 packages=setuptools.find_packages() + ['extra_requirements'],
+                 package_dir={'extra_requirements': 'extra_requirements'},
+                 package_data={'extra_requirements': ['requirements-docs.txt',
+                                                      'requirements-tests.txt']},
                  classifiers=[
-                     'Development Status :: 4 - Beta',
+                     'Development Status :: 5 - Production/Stable',
                      'License :: OSI Approved :: BSD License',
                      'Natural Language :: English',
-                     'Programming Language :: Python :: 3.6',
+                     'Programming Language :: Python :: 3.8',
+                     'Programming Language :: Python :: 3.9',
+                     'Programming Language :: Python :: 3.10',
                      'Intended Audience :: Science/Research',
                      'Programming Language :: Python :: Implementation :: CPython',
                      'Topic :: Scientific/Engineering :: Physics',
