@@ -15,7 +15,6 @@ import strax
 import straxen
 import numpy as np
 
-
 this_dir = os.path.dirname(os.path.realpath(__file__))
 
 page_header = """
@@ -23,13 +22,12 @@ page_header = """
 ===========================================================
 
 This page is an autogenerated reference for all the plugins in straxen's
-`xenonnt_online` context. 
+`{context}` context. 
 
 Colors indicate data kinds. To load tables with different data kinds,
 you currently need more than one `get_df` (or `get_array`) commands.
 
 """
-
 
 template = """
 {data_type}
@@ -38,7 +36,7 @@ template = """
 Description
 ~~~~~~~~~~~~~~~~~~~~~~
 
-Provided by plugin: {p.__class__.__name__}
+Provided by plugin: `{p.__class__.__name__} <https://github.com/XENONnT/straxen/blob/master/{module}.py>`_
 
 Data kind: {kind}
 
@@ -70,14 +68,15 @@ because changing any of those options affect this data indirectly.
 
 {config_options}
 
-
+------------------
+------------------
 """
 
-titles = {'': 'Straxen datastructure',
-          '_he': "Straxen datastructure for high energy channels",
-          '_nv': "Straxen datastructure for neutron veto",
-          '_mv': "Straxen datastructure for muon veto",
-}
+titles = {'': 'Straxen {xT} datastructure',
+          '_he': "Straxen {xT} datastructure for high energy channels",
+          '_nv': "Straxen {xT} datastructure for neutron veto",
+          '_mv': "Straxen {xT} datastructure for muon veto",
+          }
 tree_suffices = list(titles.keys())
 
 kind_colors = dict(
@@ -156,23 +155,27 @@ def get_plugins_deps(st):
     return plugins_by_deps
 
 
-def get_context():
+def get_context(is_nt):
     """
     Need to init a context without initializing the runs_db as that requires the
         appropriate passwords.
     :return: straxen context that mimics the xenonnt_online context without the rundb init
     """
-    st = straxen.contexts.xenonnt_online(_database_init=False)
-    st.context_config['forbid_creation_of'] = straxen.daqreader.DAQReader.provides
+    if is_nt:
+        st = straxen.contexts.xenonnt_online(_database_init=False, use_rucio=False)
+        st.context_config['forbid_creation_of'] = straxen.daqreader.DAQReader.provides
+    else:
+        st = straxen.contexts.xenon1t_dali()
+        st.register_all(straxen.plugins.x1t_cuts)
+        st.context_config['forbid_creation_of'] = straxen.daqreader.DAQReader.provides
     return st
 
 
-def build_datastructure_doc():
-
+def build_datastructure_doc(is_nt):
     pd.set_option('display.max_colwidth', int(1e9))
 
-    st = get_context()
-
+    st = get_context(is_nt)
+    xT = 'nT' if is_nt else '1T'
     # Too lazy to write proper graph sorter
     # Make dictionary {total number of dependencies below -> list of plugins}
 
@@ -180,9 +183,15 @@ def build_datastructure_doc():
 
     # Make graph for each suffix ('' referring to TPC)
     for suffix in tree_suffices:
-        out = page_header.format(title=titles[suffix])
-        print(f'------------ {suffix} ------------')
-        os.makedirs(this_dir + f'/graphs{suffix}', exist_ok=True)
+        title = titles[suffix].format(xT=xT)
+        out = page_header.format(title=title,
+                                 context='xenonnt_online' if is_nt else 'xenon1t_dali')
+        if not is_nt and suffix != '':
+            # No NV/MV/HE for 1T
+            continue
+
+        print(f'------------ {xT}{suffix} ------------')
+        os.makedirs(this_dir + f'/graphs{suffix}_{xT}', exist_ok=True)
         for n_deps in list(reversed(sorted(list(plugins_by_deps[suffix].keys())))):
             for data_type in plugins_by_deps[suffix][n_deps]:
                 plugins = st._get_plugins((data_type,), run_id='0')
@@ -200,12 +209,18 @@ def build_datastructure_doc():
                     for dep in p.depends_on:
                         g.edge(d, dep)
 
-                fn = this_dir + f'/graphs{suffix}/' + data_type
+                fn = this_dir + f'/graphs{suffix}_{xT}/' + data_type
                 g.render(fn)
-                with open(fn + '.svg', mode='r') as f:
+                with open(f'{fn}.svg', mode='r') as f:
                     svg = add_spaces(f.readlines()[5:])
 
-                config_df = st.show_config(d).sort_values(by='option')
+                config_df = st.show_config(data_type).sort_values(by='option')
+                # Filter out the config options of lower level datatypes
+                config_mask = []
+                for ap_to in config_df['applies_to'].values:
+                    config_mask.append(any([data_type in a for a in ap_to]))
+                keep_cols = ['option', 'default', 'current', 'help']
+                config_df = config_df[config_mask][keep_cols]
 
                 # Shorten long default values
                 config_df['default'] = [
@@ -215,21 +230,28 @@ def build_datastructure_doc():
 
                 p = plugins[data_type]
 
-                out += template.format(p=p, svg=svg, data_type=data_type,
-                    columns=add_spaces(st.data_info(data_type).to_html(index=False)),
+                out += template.format(
+                    p=p,
+                    context='',
+                    module=str(p.__module__).replace('.', '/'),
+                    svg=svg,
+                    data_type=data_type,
+                    columns=add_spaces(
+                        st.data_info(data_type).to_html(index=False)),
                     kind=p.data_kind_for(data_type),
                     docstring=p.__doc__ if p.__doc__ else '(no plugin description)',
                     config_options=add_spaces(config_df.to_html(index=False)))
 
-        with open(this_dir + f'/reference/datastructure{suffix}.rst', mode='w') as f:
+        with open(this_dir + f'/reference/datastructure{suffix}_{xT}.rst', mode='w') as f:
             f.write(out)
 
-        shutil.rmtree(this_dir + f'/graphs{suffix}')
+        shutil.rmtree(this_dir + f'/graphs{suffix}_{xT}')
 
 
 try:
     if __name__ == '__main__':
-        build_datastructure_doc()
+        build_datastructure_doc(True)
+        build_datastructure_doc(False)
 except KeyError:
     # Whatever
     pass
