@@ -1,17 +1,6 @@
-__version__ = '0.14.5'
+__version__ = '1.2.8'
 
-from warnings import warn
-# load configuration file using utilix
-try:
-    from utilix import uconfig
-# if no utilix config, get a RuntimeError
-# we don't want this to break straxen, but it does print a warning statement
-except (FileNotFoundError, RuntimeError) as e:
-    warn('Warning, utilix config file not loaded properly. copy '
-         '/project2/lgrandi/xenonnt/.xenon_config to your HOME directory!',
-         UserWarning)
-    uconfig = None
-
+from utilix import uconfig
 from .common import *
 # contexts.py below
 from .corrections_services import *
@@ -21,18 +10,28 @@ from .itp_map import *
 from .matplotlib_utils import *
 from .mini_analysis import *
 from .misc import *
-from .mongo_storage import *
-from .online_monitor import *
-from .rundb import *
+
 from .scada import *
 from .bokeh_utils import *
-
+from .url_config import *
 
 from . import plugins
 from .plugins import *
+
+from . import storage
+from .storage import *
 
 from . import analyses
 
 # Do not make all contexts directly available under straxen.
 # Otherwise we have straxen.demo() etc.
 from . import contexts
+
+from . import test_utils
+from .test_utils import *
+
+try:
+    from . import holoviews_utils
+    from .holoviews_utils import *
+except ModuleNotFoundError:
+    pass
