@@ -1,7 +1,9 @@
+from . import bokeh_waveform_plot
+from . import daq_waveforms
+from . import event_display
+from . import holoviews_waveform_display
+from . import posrec_comparison
+from . import pulse_plots
 from . import quick_checks
 from . import records_matrix
 from . import waveform_plot
-from . import holoviews_waveform_display
-from . import bokeh_waveform_plot
-from . import event_display
-from . import pulse_plots
