@@ -50,19 +50,19 @@ def mini_analysis(requires=tuple(),
 
             # Say magic words to enable holoviews
             if hv_bokeh:
+                # Generally using globals is not great, but it would be
+                # the same as doing a slow import on the top of this file
+                # pylint: disable=global-statement
                 global _hv_bokeh_initialized
                 if not _hv_bokeh_initialized:
                     import holoviews
                     holoviews.extension('bokeh')
                     _hv_bokeh_initialized = True
 
-            # TODO: This is a placeholder until the corrections system
-            # is more fully developed
             if 'to_pe' in parameters and 'to_pe' not in kwargs:
-                kwargs['to_pe'] = straxen.get_to_pe(
+                kwargs['to_pe'] = straxen.get_correction_from_cmt(
                     run_id,
-                    context.config['gain_model'],
-                    context.config['n_tpc_pmts'])
+                    context.config['gain_model'])
 
             # Prepare selection arguments
             kwargs['time_range'] = context.to_absolute_time_range(
@@ -97,7 +97,7 @@ def mini_analysis(requires=tuple(),
                 for dkind, dtypes in deps_by_kind.items():
                     if dkind in kwargs:
                         # Already have data, just apply cuts
-                        kwargs[dkind] = context.apply_selection(
+                        kwargs[dkind] = strax.apply_selection(
                             kwargs[dkind],
                             selection_str=kwargs['selection_str'],
                             time_range=kwargs['time_range'],
