@@ -6,7 +6,10 @@ from immutabledict import immutabledict
 import strax
 import straxen
 
+from straxen.get_corrections import is_cmt_option
+
 export, __all__ = strax.exporter()
+
 
 @export
 @strax.takes_config(
@@ -16,15 +19,27 @@ export, __all__ = strax.exporter()
                  help="Pretrigger time before coincidence window in ns."),
     strax.Option('resolving_time_recorder_nv', type=int, default=600,
                  help="Resolving time of the coincidence in ns."),
-    strax.Option('nbaseline_samples_lone_records_nv', type=int, default=10, track=False,
+    strax.Option('baseline_samples_nv', infer_type=False,
+                 default=('baseline_samples_nv', 'ONLINE', True), track=True,
                  help="Number of samples used in baseline rms calculation"),
+    strax.Option(
+        'hit_min_amplitude_nv', infer_type=False,
+        default=('hit_thresholds_nv', 'ONLINE', True), track=True,
+        help='Minimum hit amplitude in ADC counts above baseline. '
+             'Specify as a tuple of length n_nveto_pmts, or a number, '
+             'or a string like "pmt_commissioning_initial" which means calling '
+             'hitfinder_thresholds.py, '
+             'or a tuple like (correction=str, version=str, nT=boolean), '
+             'which means we are using cmt.'),
     strax.Option('n_lone_records_nv', type=int, default=2, track=False,
                  help="Number of lone hits to be stored per channel for diagnostic reasons."),
-    strax.Option('n_nveto_pmts', type=int, track=False,
-                 help='Number of nVETO PMTs'),
     strax.Option('channel_map', track=False, type=immutabledict,
                  help="frozendict mapping subdetector to (min, max) "
                       "channel number."),
+    strax.Option('check_raw_record_overlaps_nv',
+                 default=True, track=False, infer_type=False,
+                 help='Crash if any of the pulses in raw_records overlap with others '
+                      'in the same channel'),
 )
 class nVETORecorder(strax.Plugin):
     """
@@ -36,12 +51,16 @@ class nVETORecorder(strax.Plugin):
     properties for monitoring purposes. Depending on the setting also
     a fixed number of the lone_records per channel are stored.
     """
-    __version__ = '0.0.6'
+    __version__ = '0.0.7'
     parallel = 'process'
 
     rechunk_on_save = True
-    save_when = strax.SaveWhen.TARGET
-    compressor = 'lz4'
+    save_when = immutabledict(
+        raw_records_coin_nv=strax.SaveWhen.TARGET,
+        lone_raw_records_nv=strax.SaveWhen.TARGET,
+        lone_raw_record_statistics_nv=strax.SaveWhen.ALWAYS,
+    )
+    compressor = 'zstd'
 
     depends_on = 'raw_records_nv'
 
@@ -51,13 +70,34 @@ class nVETORecorder(strax.Plugin):
 
     data_kind = {key: key for key in provides}
 
+    def setup(self):
+        if isinstance(self.config['baseline_samples_nv'], int):
+            self.baseline_samples = self.config['baseline_samples_nv']
+        else:
+            self.baseline_samples = straxen.get_correction_from_cmt(
+                self.run_id, self.config['baseline_samples_nv'])
+
+        # Check config of `hit_min_amplitude_nv` and define hit thresholds
+        # if cmt config
+        if is_cmt_option(self.config['hit_min_amplitude_nv']):
+            self.hit_thresholds = straxen.get_correction_from_cmt(self.run_id,
+                self.config['hit_min_amplitude_nv'])
+        # if hitfinder_thresholds config
+        elif isinstance(self.config['hit_min_amplitude_nv'], str):
+            self.hit_thresholds = straxen.hit_min_amplitude(
+                self.config['hit_min_amplitude_nv'])
+        else: # int or array
+            self.hit_thresholds = self.config['hit_min_amplitude_nv']
+
     def infer_dtype(self):
         self.record_length = strax.record_length_from_dtype(
             self.deps['raw_records_nv'].dtype_for('raw_records_nv'))
 
+        channel_range = self.config['channel_map']['nveto']
+        n_channel = (channel_range[1] - channel_range[0]) + 1
         nveto_records_dtype = strax.raw_record_dtype(self.record_length)
         nveto_diagnostic_lone_records_dtype = strax.record_dtype(self.record_length)
-        nveto_lone_records_statistics_dtype = lone_record_statistics_dtype(self.config['n_nveto_pmts'])
+        nveto_lone_records_statistics_dtype = lone_record_statistics_dtype(n_channel)
 
         dtypes = [nveto_records_dtype,
                   nveto_diagnostic_lone_records_dtype,
@@ -66,6 +106,8 @@ class nVETORecorder(strax.Plugin):
         return {k: v for k, v in zip(self.provides, dtypes)}
 
     def compute(self, raw_records_nv, start, end):
+        if self.config['check_raw_record_overlaps_nv']:
+            straxen.check_overlaps(raw_records_nv, n_channels=3000)
         # Cover the case if we do not want to have any coincidence:
         if self.config['coincidence_level_recorder_nv'] <= 1:
             rr = raw_records_nv
@@ -74,28 +116,48 @@ class nVETORecorder(strax.Plugin):
             return {'raw_records_coin_nv': rr,
                     'lone_raw_records_nv': lr,
                     'lone_raw_record_statistics_nv': lrs}
-            
+
+        # Search for hits to define coincidence intervals:
+        temp_records = strax.raw_to_records(raw_records_nv)
+        temp_records = strax.sort_by_time(temp_records)
+        strax.zero_out_of_bounds(temp_records)
+        strax.baseline(temp_records,
+                       baseline_samples=self.baseline_samples,
+                       flip=True)
+        hits = strax.find_hits(temp_records, min_amplitude=self.hit_thresholds)
+        del temp_records
+
         # First we have to split rr into records and lone records:
         # Please note that we consider everything as a lone record which
         # does not satisfy the coincidence requirement
-        intervals = coincidence(raw_records_nv,
-                                self.config['coincidence_level_recorder_nv'],
-                                self.config['resolving_time_recorder_nv'],
-                                self.config['pre_trigger_time_nv']
-                               )
+        intervals = find_coincidence(hits,
+                                     self.config['coincidence_level_recorder_nv'],
+                                     self.config['resolving_time_recorder_nv'],
+                                     self.config['pre_trigger_time_nv'])
+        del hits
+
         # Always save the first and last resolving_time nanoseconds (e.g. 600 ns)  since we cannot guarantee the gap
         # size to be larger. (We cannot use an OverlapingWindow plugin either since it requires disjoint objects.)
         if len(intervals):
-            intervals_with_bounds = np.zeros((len(intervals) + 2, 2), dtype=np.int64)
-            intervals_with_bounds[1:-1, :] = intervals
-            intervals_with_bounds[0, :] = start, min(start + self.config['resolving_time_recorder_nv'], intervals[0, 0])
-            intervals_with_bounds[-1, :] = max(end - self.config['resolving_time_recorder_nv'], intervals[-1, 1]), end
+            intervals_with_bounds = np.zeros(len(intervals) + 2, dtype=strax.time_fields)
+            intervals_with_bounds['time'][1:-1] = intervals['time']
+            intervals_with_bounds['endtime'][1:-1] = intervals['endtime']
+            intervals_with_bounds['time'][0] = start
+            intervals_with_bounds['endtime'][0] = min(start + self.config['resolving_time_recorder_nv'],
+                                                      intervals['time'][0])
+            intervals_with_bounds['time'][-1] = max(end - self.config['resolving_time_recorder_nv'],
+                                                    intervals['endtime'][-1])
+            intervals_with_bounds['endtime'][-1] = end
             del intervals
         else:
-            intervals_with_bounds = np.zeros((0, 2), dtype=np.int64)
+            intervals_with_bounds = np.zeros((0, 2), dtype=strax.time_fields)
 
         neighbors = strax.record_links(raw_records_nv)
-        mask = pulse_in_interval(raw_records_nv, neighbors, *np.transpose(intervals_with_bounds))
+        mask = pulse_in_interval(raw_records_nv,
+                                 neighbors,
+                                 intervals_with_bounds['time'],
+                                 intervals_with_bounds['endtime'],)
+
         rr, lone_records = straxen.mask_and_not(raw_records_nv, mask)
 
         # Compute some properties of the lone_records:
@@ -108,7 +170,7 @@ class nVETORecorder(strax.Plugin):
         lr = strax.sort_by_time(lr)
         strax.zero_out_of_bounds(lr)
         strax.baseline(lr,
-                       baseline_samples=self.config['nbaseline_samples_lone_records_nv'],
+                       baseline_samples=self.baseline_samples,
                        flip=True)
         strax.integrate(lr)
         lrs, lr = compute_lone_records(lr, self.config['channel_map']['nveto'], self.config['n_lone_records_nv'])
@@ -159,7 +221,7 @@ def compute_lone_records(lone_records, nveto_channels, n):
         The array shape is of the raw_records dtype.
     """
     ch0, ch119 = nveto_channels
-    
+
     if len(lone_records):
         # Results computation of lone records:
         res = np.zeros(1, dtype=lone_record_statistics_dtype(ch119+1-ch0))
@@ -175,7 +237,7 @@ def compute_lone_records(lone_records, nveto_channels, n):
 
 
 @numba.njit(nogil=True, cache=True)
-def _compute_lone_records(lone_record, res, lone_ids, n,  nveto_channels):
+def _compute_lone_records(lone_record, res, lone_ids, n, nveto_channels):
     ch0, ch119 = nveto_channels
     n_channels = ch119 - ch0 + 1
 
@@ -294,7 +356,7 @@ def pulse_in_interval(raw_records, record_links, start_times, end_times):
 
 
 @export
-def coincidence(records, nfold=4, resolving_time=300, pre_trigger=0):
+def find_coincidence(records, nfold=4, resolving_time=300, pre_trigger=0):
     """
     Checks if n-neighboring events are less apart from each other then
     the specified resolving time.
@@ -313,11 +375,18 @@ def coincidence(records, nfold=4, resolving_time=300, pre_trigger=0):
          they will be merged into a single interval.
     """
     if len(records):
-        start_times = _coincidence(records, nfold, resolving_time)
-        intervals = _merge_intervals(start_times-pre_trigger, 
-                                     resolving_time+pre_trigger)
+        if nfold > 1:
+            start_times = _coincidence(records, nfold, resolving_time)
+        else:
+            # In case of a "single-fold" coincidence every thing gives
+            # the start of a new interval:
+            start_times = records['time']
+        intervals = np.zeros(len(start_times), dtype=strax.time_fields)
+        intervals['time'] = start_times - pre_trigger
+        intervals['endtime'] = start_times + resolving_time
+        intervals = merge_intervals(intervals)
     else:
-        intervals = np.zeros((0, 2), np.int64)
+        intervals = np.zeros(0, dtype=strax.time_fields)
     return intervals
 
 
@@ -340,7 +409,7 @@ def _coincidence(rr, nfold=4, resolving_time=300):
     t_diff = np.diff(start_times, prepend=start_times[0])
 
     # 2. Now we have to check if n-events are within resolving time:
-    #   -> use moving average with size n to accumulate time between n-pulses
+    #   -> use moving sum with size n to accumulate time between n-pulses
     #   -> check if accumulated time is below resolving time
 
     # generate kernel:
@@ -352,34 +421,52 @@ def _coincidence(rr, nfold=4, resolving_time=300):
     # Do not have to check the last n-1 events since by definition they can not satisfy the n-fold coincidence.
     # So we can keep the mask false.
     t_cum = t_cum[:-(nfold - 1)]
-    mask[:-(nfold - 1)] = t_cum <= resolving_time
+    mask[:-(nfold - 1)] = t_cum < resolving_time
     return start_times[mask]
 
 
-@numba.njit(nogil=True, cache=True)
-def _merge_intervals(start_time, resolving_time):
+@export
+def merge_intervals(intervals):
     """
-    Function which merges overlapping time intervals into a single one.
+    Function which merges overlapping intervals into a single one.
 
-    Note:
-        If start times of two intervals are exactly resolving_time apart
-        from each other they will be merged into a single interval.
+    :param intervals: Any numpy array with strax time fields.
+    :returns: New intervals with time and endtime according to the
+        overlapping intervals.
     """
-    # check for gaps larger than resolving_time:
-    # The gaps will indicate the starts of new intervals
-    gaps = np.diff(start_time) > resolving_time
-    
-    last_element = np.argwhere(gaps).flatten()
-    first_element = 0
-    # Creating output
-    # There is one more interval than gaps
-    intervals = np.zeros((np.sum(gaps) + 1, 2), dtype=np.int64)
+    res = np.zeros(len(intervals), dtype=strax.time_fields)
 
-    # Looping over all intervals, except for the last one:
-    for ind, le in enumerate(last_element):
-        intervals[ind] = (start_time[first_element], start_time[le]+resolving_time)
-        first_element = le + 1
+    if len(intervals):
+        res = _merge_intervals(intervals['time'],
+                               strax.endtime(intervals),
+                               res)
+    return res
 
-    # Now we have to deal with the last gap:
-    intervals[-1] = (start_time[first_element], start_time[first_element:][-1] + resolving_time)
-    return intervals
+
+@numba.njit(cache=True, nogil=True)
+def _merge_intervals(start, end, res):
+    offset = 0
+
+    interval_start = start[0]
+    interval_end = end[0]
+    for next_interval_start, next_interval_end in zip(start[1:], end[1:]):
+        if interval_end >= next_interval_start:
+            # Interval overlaps, updated only end:
+            interval_end = next_interval_end
+            continue
+
+        # Intervals do not overlap, save interval:
+        res[offset]['time'] = interval_start
+        res[offset]['endtime'] = interval_end
+        offset += 1
+
+        # New interval:
+        interval_start = next_interval_start
+        interval_end = next_interval_end
+
+    # Save last interval:
+    res[offset]['time'] = interval_start
+    res[offset]['endtime'] = interval_end
+    offset += 1
+
+    return res[:offset]
