@@ -1,24 +1,26 @@
 import numba
 import numpy as np
-
 import strax
+from immutabledict import immutabledict
+from strax.processing.general import _touching_windows
 import straxen
 from .pulse_processing import HITFINDER_OPTIONS, HITFINDER_OPTIONS_he, HE_PREAMBLE
-from strax.processing.general import _touching_windows
-from warnings import warn
+from straxen.get_corrections import is_cmt_option
+
 
 export, __all__ = strax.exporter()
+FAKE_MERGED_S2_TYPE = -42
 
 
 @export
 @strax.takes_config(
-    strax.Option('peaklet_gap_threshold', default=350,
+    strax.Option('peaklet_gap_threshold', default=700, infer_type=False,
                  help="No hits for this many ns triggers a new peak"),
-    strax.Option('peak_left_extension', default=30,
+    strax.Option('peak_left_extension', default=30, infer_type=False,
                  help="Include this many ns left of hits in peaks"),
-    strax.Option('peak_right_extension', default=200,
+    strax.Option('peak_right_extension', default=200, infer_type=False,
                  help="Include this many ns right of hits in peaks"),
-    strax.Option('peak_min_pmts', default=2,
+    strax.Option('peak_min_pmts', default=2, infer_type=False,
                  help="Minimum number of contributing PMTs needed to define a peak"),
     strax.Option('peak_split_gof_threshold',
                  # See https://xe1t-wiki.lngs.infn.it/doku.php?id=
@@ -27,39 +29,44 @@ export, __all__ = strax.exporter()
                  # for more information
                  default=(
                      None,  # Reserved
-                     ((0.5, 1), (4, 0.4)),
-                     ((2, 1), (4.5, 0.4))),
+                     ((0.5, 1.0), (6.0, 0.4)),
+                     ((2.5, 1.0), (5.625, 0.4))), infer_type=False,
                  help='Natural breaks goodness of fit/split threshold to split '
                       'a peak. Specify as tuples of (log10(area), threshold).'),
-    strax.Option('peak_split_filter_wing_width', default=70,
+    strax.Option('peak_split_filter_wing_width', default=70, infer_type=False,
                  help='Wing width of moving average filter for '
                       'low-split natural breaks'),
-    strax.Option('peak_split_min_area', default=40.,
+    strax.Option('peak_split_min_area', default=40., infer_type=False,
                  help='Minimum area to evaluate natural breaks criterion. '
                       'Smaller peaks are not split.'),
-    strax.Option('peak_split_iterations', default=20,
+    strax.Option('peak_split_iterations', default=20, infer_type=False,
                  help='Maximum number of recursive peak splits to do.'),
-    strax.Option('diagnose_sorting', track=False, default=False,
+    strax.Option('diagnose_sorting', track=False, default=False, infer_type=False,
                  help="Enable runtime checks for sorting and disjointness"),
-    strax.Option('gain_model',
-                 help='PMT gain model. Specify as (model_type, model_config)'),
-    strax.Option('tight_coincidence_window_left', default=50,
+    strax.Option('gain_model', infer_type=False,
+                 help='PMT gain model. Specify as '
+                 '(str(model_config), str(version), nT-->boolean'),
+    strax.Option('tight_coincidence_window_left', default=50, infer_type=False,
                  help="Time range left of peak center to call "
                       "a hit a tight coincidence (ns)"),
-    strax.Option('tight_coincidence_window_right', default=50,
+    strax.Option('tight_coincidence_window_right', default=50, infer_type=False,
                  help="Time range right of peak center to call "
                       "a hit a tight coincidence (ns)"),
     strax.Option('n_tpc_pmts', type=int,
                  help='Number of TPC PMTs'),
-    strax.Option('saturation_correction_on', default=True,
+    strax.Option('saturation_correction_on', default=True, infer_type=False,
                  help='On off switch for saturation correction'),
-    strax.Option('saturation_reference_length', default=100,
+    strax.Option('saturation_reference_length', default=100, infer_type=False,
                  help="Maximum number of reference sample used "
                       "to correct saturated samples"),
-    strax.Option('saturation_min_reference_length', default=20,
+    strax.Option('saturation_min_reference_length', default=20, infer_type=False,
                  help="Minimum number of reference sample used "
                       "to correct saturated samples"),
-
+    strax.Option('peaklet_max_duration', default=int(10e6), infer_type=False,
+                 help="Maximum duration [ns] of a peaklet"),
+    strax.Option('channel_map', track=False, type=immutabledict,
+                 help="immutabledict mapping subdetector to (min, max) "
+                      "channel number."),
     *HITFINDER_OPTIONS,
 )
 class Peaklets(strax.Plugin):
@@ -89,7 +96,7 @@ class Peaklets(strax.Plugin):
     parallel = 'process'
     compressor = 'zstd'
 
-    __version__ = '0.3.7'
+    __version__ = '0.6.0'
 
     def infer_dtype(self):
         return dict(peaklets=strax.peak_dtype(
@@ -98,22 +105,33 @@ class Peaklets(strax.Plugin):
 
     def setup(self):
         if self.config['peak_min_pmts'] > 2:
-            # Can fix by resplitting, NotImplemented
-            raise warn(f"Raising the peak_min_pmts to "
-                       f"{self.config['peak_min_pmts']} interferes with "
-                       f"lone_hit definition. See "
-                       f"github.com/XENONnT/straxen/issues/295")
-        self.to_pe = straxen.get_to_pe(self.run_id,
-                                       self.config['gain_model'],
-                                       self.config['n_tpc_pmts'])
+            # Can fix by re-splitting,
+            raise NotImplementedError(
+                f"Raising the peak_min_pmts to {self.config['peak_min_pmts']} "
+                f"interferes with lone_hit definition. "
+                f"See github.com/XENONnT/straxen/issues/295")
+
+        self.to_pe = straxen.get_correction_from_cmt(self.run_id,
+                                       self.config['gain_model'])
+
+        # Check config of `hit_min_amplitude` and define hit thresholds
+        # if cmt config
+        if is_cmt_option(self.config['hit_min_amplitude']):
+            self.hit_thresholds = straxen.get_correction_from_cmt(self.run_id,
+                self.config['hit_min_amplitude'])
+        # if hitfinder_thresholds config
+        elif isinstance(self.config['hit_min_amplitude'], str):
+            self.hit_thresholds = straxen.hit_min_amplitude(
+                self.config['hit_min_amplitude'])
+        else: # int or array
+            self.hit_thresholds = self.config['hit_min_amplitude']
+            
+        self.channel_range = self.config['channel_map']['tpc']
 
     def compute(self, records, start, end):
         r = records
 
-        hits = strax.find_hits(
-            r,
-            min_amplitude=straxen.hit_min_amplitude(
-                self.config['hit_min_amplitude']))
+        hits = strax.find_hits(r, min_amplitude=self.hit_thresholds)
 
         # Remove hits in zero-gain channels
         # they should not affect the clustering!
@@ -129,7 +147,9 @@ class Peaklets(strax.Plugin):
             left_extension=self.config['peak_left_extension'],
             right_extension=self.config['peak_right_extension'],
             min_channels=self.config['peak_min_pmts'],
-            result_dtype=self.dtype_for('peaklets'))
+            result_dtype=self.dtype_for('peaklets'),
+            max_duration=self.config['peaklet_max_duration'],
+        )
 
         # Make sure peaklets don't extend out of the chunk boundary
         # This should be very rare in normal data due to the ADC pretrigger
@@ -139,7 +159,8 @@ class Peaklets(strax.Plugin):
         # Get hits outside peaklets, and store them separately.
         # fully_contained is OK provided gap_threshold > extension,
         # which is asserted inside strax.find_peaks.
-        lone_hits = hits[strax.fully_contained_in(hits, peaklets) == -1]
+        is_lone_hit = strax.fully_contained_in(hits, peaklets) == -1
+        lone_hits = hits[is_lone_hit]
         strax.integrate_lone_hits(
             lone_hits, records, peaklets,
             save_outside_hits=(self.config['peak_left_extension'],
@@ -147,14 +168,40 @@ class Peaklets(strax.Plugin):
             n_channels=len(self.to_pe))
 
         # Compute basic peak properties -- needed before natural breaks
-        strax.sum_waveform(peaklets, r, self.to_pe)
+        hits = hits[~is_lone_hit]
+        # Define regions outside of peaks such that _find_hit_integration_bounds
+        # is not extended beyond a peak.
+        outside_peaks = self.create_outside_peaks_region(peaklets, start, end)
+        strax.find_hit_integration_bounds(
+            hits, outside_peaks, records,
+            save_outside_hits=(self.config['peak_left_extension'],
+                               self.config['peak_right_extension']),
+            n_channels=len(self.to_pe),
+            allow_bounds_beyond_records=True,
+        )
+
+        # Transform hits to hitlets for naming conventions. A hit refers
+        # to the central part above threshold a hitlet to the entire signal
+        # including the left and right extension.
+        # (We are not going to use the actual hitlet data_type here.)
+        hitlets = hits
+        del hits
+
+        hitlet_time_shift = (hitlets['left'] - hitlets['left_integration']) * hitlets['dt']
+        hitlets['time'] = hitlets['time'] - hitlet_time_shift
+        hitlets['length'] = (hitlets['right_integration'] - hitlets['left_integration'])
+        hitlets = strax.sort_by_time(hitlets)
+        rlinks = strax.record_links(records)
+
+        strax.sum_waveform(peaklets, hitlets, r, rlinks, self.to_pe)
+
         strax.compute_widths(peaklets)
 
         # Split peaks using low-split natural breaks;
         # see https://github.com/XENONnT/straxen/pull/45
         # and https://github.com/AxFoundation/strax/pull/225
         peaklets = strax.split_peaks(
-            peaklets, r, self.to_pe,
+            peaklets, hitlets, r, rlinks, self.to_pe,
             algorithm='natural_breaks',
             threshold=self.natural_breaks_threshold,
             split_low=True,
@@ -165,11 +212,20 @@ class Peaklets(strax.Plugin):
         # Saturation correction using non-saturated channels
         # similar method used in pax
         # see https://github.com/XENON1T/pax/pull/712
+        # Cases when records is not writeable for unclear reason
+        # only see this when loading 1T test data
+        # more details on https://numpy.org/doc/stable/reference/generated/numpy.ndarray.flags.html
+        if not r['data'].flags.writeable:
+            r = r.copy()
+
         if self.config['saturation_correction_on']:
-            peak_saturation_correction(
-                r, peaklets, self.to_pe,
+            peak_list = peak_saturation_correction(
+                r, rlinks, peaklets, hitlets, self.to_pe,
                 reference_length=self.config['saturation_reference_length'],
                 min_reference_length=self.config['saturation_min_reference_length'])
+
+            # Compute the width again for corrected peaks
+            strax.compute_widths(peaklets, select_peaks_indices=peak_list)
 
         # Compute tight coincidence level.
         # Making this a separate plugin would
@@ -177,34 +233,38 @@ class Peaklets(strax.Plugin):
         # (b) increase strax memory usage / max_messages,
         #     possibly due to its currently primitive scheduling.
         hit_max_times = np.sort(
-            hits['time']
-            + hits['dt'] * hit_max_sample(records, hits))
+            hitlets['time']
+            + hitlets['dt'] * hit_max_sample(records, hitlets)
+            + hitlet_time_shift  # add time shift again to get correct maximum
+        )
         peaklet_max_times = (
                 peaklets['time']
                 + np.argmax(peaklets['data'], axis=1) * peaklets['dt'])
-        peaklets['tight_coincidence'] = get_tight_coin(
+        tight_coincidence_channel = get_tight_coin(
             hit_max_times,
+            hitlets['channel'],
             peaklet_max_times,
             self.config['tight_coincidence_window_left'],
-            self.config['tight_coincidence_window_right'])
+            self.config['tight_coincidence_window_right'],
+            self.channel_range)
+
+        peaklets['tight_coincidence'] = tight_coincidence_channel
 
         if self.config['diagnose_sorting'] and len(r):
             assert np.diff(r['time']).min(initial=1) >= 0, "Records not sorted"
-            assert np.diff(hits['time']).min(initial=1) >= 0, "Hits not sorted"
+            assert np.diff(hitlets['time']).min(initial=1) >= 0, "Hits/Hitlets not sorted"
             assert np.all(peaklets['time'][1:]
                           >= strax.endtime(peaklets)[:-1]), "Peaks not disjoint"
-        
+
         # Update nhits of peaklets:
-        counts = strax.touching_windows(hits, peaklets)
+        counts = strax.touching_windows(hitlets, peaklets)
         counts = np.diff(counts, axis=1).flatten()
-        counts += 1
         peaklets['n_hits'] = counts
-        
+
         return dict(peaklets=peaklets,
                     lone_hits=lone_hits)
 
     def natural_breaks_threshold(self, peaks):
-        # TODO avoid duplication with PeakBasics somehow?
         rise_time = -peaks['area_decile_from_midpoint'][:, 1]
 
         # This is ~1 for an clean S2, ~0 for a clean S1,
@@ -231,16 +291,43 @@ class Peaklets(strax.Plugin):
             if strax.endtime(p) > end:
                 p['length'] = (end - p['time']) // p['dt']
 
+    @staticmethod
+    def create_outside_peaks_region(peaklets, start, end):
+        """
+        Creates time intervals which are outside peaks.
+
+        :param peaklets: Peaklets for which intervals should be computed.
+        :param start: Chunk start
+        :param end: Chunk end
+        :return: array of strax.time_fields dtype.
+        """
+        if not len(peaklets):
+            return np.zeros(0, dtype=strax.time_fields)
+        
+        outside_peaks = np.zeros(len(peaklets) + 1,
+                                 dtype=strax.time_fields)
+        
+        outside_peaks[0]['time'] = start
+        outside_peaks[0]['endtime'] = peaklets[0]['time']
+        outside_peaks[1:-1]['time'] = strax.endtime(peaklets[:-1])
+        outside_peaks[1:-1]['endtime'] = peaklets['time'][1:]
+        outside_peaks[-1]['time'] = strax.endtime(peaklets[-1])
+        outside_peaks[-1]['endtime'] = end
+        return outside_peaks
+
 
 @numba.jit(nopython=True, nogil=True, cache=True)
-def peak_saturation_correction(records, peaks, to_pe,
+def peak_saturation_correction(records, rlinks, peaks, hitlets, to_pe,
                                reference_length=100,
                                min_reference_length=20,
                                use_classification=False,
                                ):
     """Correct the area and per pmt area of peaks from saturation
     :param records: Records
+    :param rlinks: strax.record_links of corresponding records.
     :param peaks: Peaklets / Peaks
+    :param hitlets: Hitlets found in records to build peaks.
+        (Hitlets are hits including the left/right extension)
     :param to_pe: adc to PE conversion (length should equal number of PMTs)
     :param reference_length: Maximum number of reference sample used
     to correct saturated samples
@@ -312,7 +399,8 @@ def peak_saturation_correction(records, peaks, to_pe,
         peaks[peak_i]['length'] = p['length'] * p['dt'] / dt
         peaks[peak_i]['dt'] = dt
 
-    strax.sum_waveform(peaks, records, to_pe, peak_list)
+    strax.sum_waveform(peaks, hitlets, records, rlinks, to_pe, peak_list)
+    return peak_list
 
 
 @numba.jit(nopython=True, nogil=True, cache=True)
@@ -339,7 +427,13 @@ def _peak_saturation_correction_inner(channel_saturated, records, p,
 
         # Define the reference region as reference_length before the first saturation point
         # unless there are not enough samples
-        s0 = np.argmax(b >= r0['baseline'])
+        bl = np.inf
+        for record_i in b_index[ch]:
+            if record_i == -1:
+                break
+            bl = min(bl, records['baseline'][record_i])
+
+        s0 = np.argmax(b >= np.int16(bl))
         ref = slice(max(0, s0-reference_length), s0)
 
         if (b[ref] * to_pe[ch] > 1).sum() < min_reference_length:
@@ -349,7 +443,7 @@ def _peak_saturation_correction_inner(channel_saturated, records, p,
         if (b_sumwf[ref] > 1).sum() < min_reference_length:
             # the same condition applies to the waveform model
             continue
-        if np.sum(b[ref]) * to_pe[ch] / np.sum(b_sumwf[ref]) < 1:
+        if np.sum(b[ref]) * to_pe[ch] / np.sum(b_sumwf[ref]) > 1:
             # The pulse is saturated, but insufficient information is available in the other channels
             # to reliably reconstruct it
             continue
@@ -358,6 +452,8 @@ def _peak_saturation_correction_inner(channel_saturated, records, p,
 
         # Loop over the record indices of the saturated channel (saved in b_index buffer)
         for record_i in b_index[ch]:
+            if record_i == -1:
+                break
             r = records[record_i]
             r_slice, b_slice = strax.overlap_indices(
                 r['time'] // dt, r['length'],
@@ -387,18 +483,18 @@ def _peak_saturation_correction_inner(channel_saturated, records, p,
 
 @export
 @strax.takes_config(
-    strax.Option('n_he_pmts', track=False, default=752,
+    strax.Option('n_he_pmts', track=False, default=752, infer_type=False,
                  help="Maximum channel of the he channels"),
-    strax.Option('he_channel_offset', track=False, default=500,
+    strax.Option('he_channel_offset', track=False, default=500, infer_type=False,
                  help="Minimum channel number of the he channels"),
-    strax.Option('le_to_he_amplification', default=20, track=True,
+    strax.Option('le_to_he_amplification', default=20, track=True, infer_type=False,
                  help="Difference in amplification between low energy and high "
                       "energy channels"),
-    strax.Option('peak_min_pmts_he', default=2,
+    strax.Option('peak_min_pmts_he', default=2, infer_type=False,
                  child_option=True, parent_option_name='peak_min_pmts',
                  track=True,
                  help="Minimum number of contributing PMTs needed to define a peak"),
-    strax.Option('saturation_correction_on_he', default=False,
+    strax.Option('saturation_correction_on_he', default=False, infer_type=False,
                  child_option=True, parent_option_name='saturation_correction_on',
                  track=True,
                  help='On off switch for saturation correction for High Energy'
@@ -410,20 +506,34 @@ class PeakletsHighEnergy(Peaklets):
     depends_on = 'records_he'
     provides = 'peaklets_he'
     data_kind = 'peaklets_he'
-    __version__ = '0.0.1'
+    __version__ = '0.0.2'
     child_plugin = True
+    save_when = strax.SaveWhen.TARGET
 
     def infer_dtype(self):
         return strax.peak_dtype(n_channels=self.config['n_he_pmts'])
 
     def setup(self):
-        self.to_pe = straxen.get_to_pe(self.run_id,
-                                       self.config['gain_model'],
-                                       self.config['n_tpc_pmts'])
+        self.to_pe = straxen.get_correction_from_cmt(self.run_id,
+                                       self.config['gain_model'])
 
         buffer_pmts = np.zeros(self.config['he_channel_offset'])
         self.to_pe = np.concatenate((buffer_pmts, self.to_pe))
         self.to_pe *= self.config['le_to_he_amplification']
+
+        # Check config of `hit_min_amplitude_he` and define hit thresholds
+        # if cmt config
+        if is_cmt_option(self.config['hit_min_amplitude_he']):
+            self.hit_thresholds = straxen.get_correction_from_cmt(self.run_id,
+                self.config['hit_min_amplitude_he'])
+        # if hitfinder_thresholds config
+        elif isinstance(self.config['hit_min_amplitude_he'], str):
+            self.hit_thresholds = straxen.hit_min_amplitude(
+                self.config['hit_min_amplitude_he'])
+        else: # int or array
+            self.hit_thresholds = self.config['hit_min_amplitude_he']
+            
+        self.channel_range = self.config['channel_map']['he']
 
     def compute(self, records_he, start, end):
         result = super().compute(records_he, start, end)
@@ -432,43 +542,83 @@ class PeakletsHighEnergy(Peaklets):
 
 @export
 @strax.takes_config(
-    strax.Option('s1_max_rise_time', default=60,
-                 help="Maximum S1 rise time for < 100 PE [ns]"),
-    strax.Option('s1_max_rise_time_post100', default=150,
+    strax.Option('s1_risetime_area_parameters', default=(50, 80, 12), type=(list, tuple),
+                 help="norm, const, tau in the empirical boundary in the risetime-area plot"),
+    strax.Option('s1_risetime_aft_parameters', default=(-1, 2.6), type=(list, tuple),
+                 help=("Slope and offset in exponential of emperical boundary in the rise time-AFT "
+                      "plot. Specified as (slope, offset)")),
+    strax.Option('s1_flatten_threshold_aft', default=(0.6, 100), type=(tuple, list),
+                 help=("Threshold for AFT, above which we use a flatted boundary for rise time" 
+                       "Specified values: (AFT boundary, constant rise time).")),
+    strax.Option('n_top_pmts', default=straxen.n_top_pmts, type=int,
+                 help="Number of top PMTs"),
+    strax.Option('s1_max_rise_time_post100', default=200, type=(int, float),
                  help="Maximum S1 rise time for > 100 PE [ns]"),
-    strax.Option('s1_min_coincidence', default=3,
+    strax.Option('s1_min_coincidence', default=2, type=int,
                  help="Minimum tight coincidence necessary to make an S1"),
-    strax.Option('s2_min_pmts', default=4,
+    strax.Option('s2_min_pmts', default=4, type=int,
                  help="Minimum number of PMTs contributing to an S2"))
 class PeakletClassification(strax.Plugin):
     """Classify peaklets as unknown, S1, or S2."""
     provides = 'peaklet_classification'
     depends_on = ('peaklets',)
     parallel = True
-    dtype = (
-        strax.interval_dtype +
-        [('type', np.int8, 'Classification of the peak(let)'),])
-    __version__ = '0.2.0'
+    dtype = (strax.peak_interval_dtype
+             + [('type', np.int8, 'Classification of the peak(let)')])
+
+    __version__ = '3.0.3'
+
+    @staticmethod
+    def upper_rise_time_area_boundary(area, norm, const, tau):
+        """
+        Function which determines the upper boundary for the rise-time
+        for a given area.
+        """
+        return norm*np.exp(-area/tau) + const
+
+    @staticmethod
+    def upper_rise_time_aft_boundary(aft, slope, offset, aft_boundary, flat_threshold):
+        """
+        Function which computes the upper rise time boundary as a function
+        of area fraction top.
+        """
+        res = 10**(slope * aft + offset)
+        res[aft >= aft_boundary] = flat_threshold
+        return res
 
     def compute(self, peaklets):
-        peaks = peaklets
-
         ptype = np.zeros(len(peaklets), dtype=np.int8)
 
-        # Properties needed for classification. Bit annoying these computations
-        # are duplicated in peak_basics curently...
-        rise_time = -peaks['area_decile_from_midpoint'][:, 1]
-        n_channels = (peaks['area_per_channel'] > 0).sum(axis=1)
+        # Properties needed for classification:
+        rise_time = -peaklets['area_decile_from_midpoint'][:, 1]
+        n_channels = (peaklets['area_per_channel'] > 0).sum(axis=1)
+        n_top = self.config['n_top_pmts']
+        area_top = peaklets['area_per_channel'][:, :n_top].sum(axis=1)
+        area_total = peaklets['area_per_channel'].sum(axis=1)
+        area_fraction_top = area_top/area_total
 
-        is_s1 = (
-           (rise_time <= self.config['s1_max_rise_time'])
-            | ((rise_time <= self.config['s1_max_rise_time_post100'])
-               & (peaks['area'] > 100)))
-        is_s1 &= peaks['tight_coincidence'] >= self.config['s1_min_coincidence']
-        ptype[is_s1] = 1
+        is_large_s1 = (peaklets['area'] >= 100)
+        is_large_s1 &= (rise_time <= self.config['s1_max_rise_time_post100'])
+        is_large_s1 &= peaklets['tight_coincidence'] >= self.config['s1_min_coincidence']
+
+        is_small_s1 = peaklets["area"] < 100
+        is_small_s1 &= rise_time < self.upper_rise_time_area_boundary(
+            peaklets["area"],
+            *self.config["s1_risetime_area_parameters"],
+        )
+
+        is_small_s1 &= rise_time < self.upper_rise_time_aft_boundary(
+            area_fraction_top,
+            *self.config["s1_risetime_aft_parameters"],
+            *self.config["s1_flatten_threshold_aft"],
+        )
+
+        is_small_s1 &= peaklets['tight_coincidence'] >= self.config['s1_min_coincidence']
+
+        ptype[is_large_s1 | is_small_s1] = 1
 
         is_s2 = n_channels >= self.config['s2_min_pmts']
-        is_s2[is_s1] = False
+        is_s2[is_large_s1 | is_small_s1] = False
         ptype[is_s2] = 2
 
         return dict(type=ptype,
@@ -487,69 +637,96 @@ class PeakletClassificationHighEnergy(PeakletClassification):
     __doc__ = HE_PREAMBLE + PeakletClassification.__doc__
     provides = 'peaklet_classification_he'
     depends_on = ('peaklets_he',)
-    __version__ = '0.0.1'
+    __version__ = '0.0.2'
     child_plugin = True
 
     def compute(self, peaklets_he):
         return super().compute(peaklets_he)
-    
-
-FAKE_MERGED_S2_TYPE = -42
 
 
 @export
 @strax.takes_config(
-    strax.Option('s2_merge_max_area', default=5000.,
-                 help="Merge peaklet cluster only if area < this [PE]"),
-    strax.Option('s2_merge_max_gap', default=3500,
-                 help="Maximum separation between peaklets to allow merging [ns]"),
-    strax.Option('s2_merge_max_duration', default=15_000,
+    strax.Option('s2_merge_max_duration', default=50_000, infer_type=False,
                  help="Do not merge peaklets at all if the result would be a peak "
-                      "longer than this [ns]"))
+                      "longer than this [ns]"),
+    strax.Option('s2_merge_gap_thresholds', default=((1.7, 2.65e4), (4.0, 2.6e3), (5.0, 0.)),
+                 infer_type=False,
+                 help="Points to define maximum separation between peaklets to allow "
+                      "merging [ns] depending on log10 area of the merged peak\n"
+                      "where the gap size of the first point is the maximum gap to allow merging"
+                      "and the area of the last point is the maximum area to allow merging. "
+                      "The format is ((log10(area), max_gap), (..., ...), (..., ...))"
+                 ),
+    strax.Option('gain_model', infer_type=False,
+                 help='PMT gain model. Specify as '
+                      '(str(model_config), str(version), nT-->boolean'),
+    strax.Option('merge_without_s1', default=True, infer_type=False,
+                 help="If true, S1s will be igored during the merging. "
+                      "It's now possible for a S1 to be inside a S2 post merging"),
+)
 class MergedS2s(strax.OverlapWindowPlugin):
     """
     Merge together peaklets if peak finding favours that they would
     form a single peak instead.
     """
-    depends_on = ('peaklets', 'peaklet_classification')
+    depends_on = ('peaklets', 'peaklet_classification', 'lone_hits')
     data_kind = 'merged_s2s'
     provides = 'merged_s2s'
+    __version__ = '0.4.1'
+
+    def setup(self):
+        self.to_pe = straxen.get_correction_from_cmt(self.run_id,
+                                                     self.config['gain_model'])
 
     def infer_dtype(self):
         return strax.unpack_dtype(self.deps['peaklets'].dtype_for('peaklets'))
 
     def get_window_size(self):
-        return 5 * (self.config['s2_merge_max_gap']
+        return 5 * (int(self.config['s2_merge_gap_thresholds'][0][1])
                     + self.config['s2_merge_max_duration'])
 
-    def compute(self, peaklets):
-        if not len(peaklets):
-            return peaklets[:0]
+    def compute(self, peaklets, lone_hits):
+        if self.config['merge_without_s1']:
+            peaklets = peaklets[peaklets['type'] != 1]
 
-        if self.config['s2_merge_max_gap'] < 0:
+        if len(peaklets) <= 1:
+            return np.zeros(0, dtype=peaklets.dtype)
+
+        gap_thresholds = self.config['s2_merge_gap_thresholds']
+        max_gap = gap_thresholds[0][1]
+        max_area = 10 ** gap_thresholds[-1][0]
+
+        if max_gap < 0:
             # Do not merge at all
-            merged_s2s = np.zeros(0, dtype=peaklets.dtype)
+            return np.zeros(0, dtype=peaklets.dtype)
         else:
-            # Find all groups of peaklets separated by < the gap
-            cluster_starts, cluster_stops = strax.find_peak_groups(
-                peaklets,
-                self.config['s2_merge_max_gap'])
-
+            # Max gap and area should be set by the gap thresholds
+            # to avoid contradictions
             start_merge_at, end_merge_at = self.get_merge_instructions(
                 peaklets['time'], strax.endtime(peaklets),
                 areas=peaklets['area'],
                 types=peaklets['type'],
-                cluster_starts=cluster_starts,
-                cluster_stops=cluster_stops,
+                gap_thresholds=gap_thresholds,
                 max_duration=self.config['s2_merge_max_duration'],
-                max_area=self.config['s2_merge_max_area'])
-
+                max_gap=max_gap,
+                max_area=max_area,
+            )
             merged_s2s = strax.merge_peaks(
                 peaklets,
                 start_merge_at, end_merge_at,
-                max_buffer=int(self.config['s2_merge_max_duration']
-                               // peaklets['dt'].min()))
+                max_buffer=int(self.config['s2_merge_max_duration']//np.gcd.reduce(peaklets['dt'])),
+            )
             merged_s2s['type'] = 2
+            
+            # Updated time and length of lone_hits and sort again:
+            lh = np.copy(lone_hits)
+            del lone_hits
+            lh_time_shift = (lh['left'] - lh['left_integration']) *lh['dt']
+            lh['time'] = lh['time'] - lh_time_shift
+            lh['length'] = (lh['right_integration'] - lh['left_integration'])
+            lh = strax.sort_by_time(lh)
+            strax.add_lone_hits(merged_s2s, lh, self.to_pe)
+
             strax.compute_widths(merged_s2s)
 
         return merged_s2s
@@ -558,47 +735,93 @@ class MergedS2s(strax.OverlapWindowPlugin):
     @numba.njit(cache=True, nogil=True)
     def get_merge_instructions(
             peaklet_starts, peaklet_ends, areas, types,
-            cluster_starts, cluster_stops,
-            max_duration, max_area):
-        start_merge_at = np.zeros(len(cluster_starts), dtype=np.int32)
-        end_merge_at = np.zeros(len(cluster_starts), dtype=np.int32)
-        n_to_merge = 0
-        left_i = 0
+            gap_thresholds, max_duration, max_gap, max_area):
+        """
+        Finding the group of peaklets to merge. To do this start with the
+        smallest gaps and keep merging until the new, merged S2 has such a
+        large area or gap to adjacent peaks that merging is not required
+        anymore.
+        see https://github.com/XENONnT/straxen/pull/548 and https://github.com/XENONnT/straxen/pull/568
 
-        for cluster_i, cluster_start in enumerate(cluster_starts):
-            cluster_stop = cluster_stops[cluster_i]
+        :returns: list of the first index of peaklet to be merged and
+        list of the exclusive last index of peaklet to be merged
+        """
 
-            if cluster_stop - cluster_start > max_duration:
+        peaklet_gaps = peaklet_starts[1:] - peaklet_ends[:-1]
+        peaklet_start_index = np.arange(len(peaklet_starts))
+        peaklet_end_index = np.arange(len(peaklet_starts))
+
+        for gap_i in np.argsort(peaklet_gaps):
+            start_idx = peaklet_start_index[gap_i]
+            inclusive_end_idx = peaklet_end_index[gap_i + 1]
+            sum_area = np.sum(areas[start_idx:inclusive_end_idx + 1])
+            this_gap = peaklet_gaps[gap_i]
+
+            if inclusive_end_idx < start_idx:
+                raise ValueError('Something went wrong, left is bigger then right?!')
+
+            if this_gap > max_gap:
+                break
+            if sum_area > max_area:
+                # For very large S2s, we assume that natural breaks is taking care
+                continue
+            if (sum_area > 0) and (
+                    this_gap > merge_s2_threshold(np.log10(sum_area),
+                                                  gap_thresholds)):
+                # The merged peak would be too large
                 continue
 
-            # Recover left and right indices of the clusters.
-            # stops are inclusive for a few lines... sorry...
-            while peaklet_starts[left_i] < cluster_start:
-                left_i += 1
-            right_i = left_i
-            while peaklet_ends[right_i] < cluster_stop:
-                right_i += 1
-
-            if left_i == right_i:
-                # One peak, nothing to merge
+            peak_duration = (peaklet_ends[inclusive_end_idx] - peaklet_starts[start_idx])
+            if peak_duration >= max_duration:
                 continue
 
-            if types[left_i] != 2:
-                # Doesn't start with S2: do not merge
-                continue
+            # Merge gap in other words this means p @ gap_i and p @gap_i + 1 share the same
+            # start, end and area:
+            peaklet_start_index[start_idx:inclusive_end_idx + 1] = peaklet_start_index[start_idx]
+            peaklet_end_index[start_idx:inclusive_end_idx + 1] = peaklet_end_index[inclusive_end_idx]
 
-            right_i += 1   # From here on, right_i is exclusive
+        start_merge_at = np.unique(peaklet_start_index)
+        end_merge_at = np.unique(peaklet_end_index)
+        if not len(start_merge_at) == len(end_merge_at):
+            raise ValueError('inconsistent start and end merge instructions')
 
-            if areas[left_i:right_i].sum() > max_area:
-                continue
+        merge_start, merge_stop_exclusive = _filter_s1_starts(
+            start_merge_at, types, end_merge_at)
 
-            start_merge_at[n_to_merge] = left_i
-            end_merge_at[n_to_merge] = right_i
-            n_to_merge += 1
+        return merge_start, merge_stop_exclusive
 
-        return start_merge_at[:n_to_merge], end_merge_at[:n_to_merge]
 
-    
+@numba.njit(cache=True, nogil=True)
+def _filter_s1_starts(start_merge_at, types, end_merge_at):
+    for start_merge_idx, _ in enumerate(start_merge_at):
+        while types[start_merge_at[start_merge_idx]] != 2:
+            if end_merge_at[start_merge_idx] - start_merge_at[start_merge_idx] <= 1:
+                break
+            start_merge_at[start_merge_idx] += 1
+
+    start_merge_with_s2 = types[start_merge_at] == 2
+    merges_at_least_two_peaks = end_merge_at - start_merge_at >= 1
+
+    keep_merges = start_merge_with_s2 & merges_at_least_two_peaks
+    return start_merge_at[keep_merges], end_merge_at[keep_merges] + 1
+
+
+@numba.njit(cache=True, nogil=True)
+def merge_s2_threshold(log_area, gap_thresholds):
+    """Return gap threshold for log_area of the merged S2
+    with linear interpolation given the points in gap_thresholds
+    :param log_area: Log 10 area of the merged S2
+    :param gap_thresholds: tuple (n, 2) of fix points for interpolation
+    """
+    for i, (a1, g1) in enumerate(gap_thresholds):
+        if log_area < a1:
+            if i == 0:
+                return g1
+            a0, g0 = gap_thresholds[i - 1]
+            return (log_area - a0) * (g1 - g0) / (a1 - a0) + g0
+    return gap_thresholds[-1][1]
+
+
 @export
 class MergedS2sHighEnergy(MergedS2s):
     __doc__ = HE_PREAMBLE + MergedS2s.__doc__
@@ -612,13 +835,20 @@ class MergedS2sHighEnergy(MergedS2s):
         return strax.unpack_dtype(self.deps['peaklets_he'].dtype_for('peaklets_he'))
 
     def compute(self, peaklets_he):
-        return super().compute(peaklets_he)
+        # There are not any lone hits for the high energy channel, 
+        #  so create a dummy for the compute method.
+        lone_hits = np.zeros(0, dtype=strax.hit_dtype)
+        return super().compute(peaklets_he, lone_hits)
 
 
 @export
 @strax.takes_config(
-    strax.Option('diagnose_sorting', track=False, default=False,
-                 help="Enable runtime checks for sorting and disjointness"))
+    strax.Option('diagnose_sorting', track=False, default=False, infer_type=False,
+                 help="Enable runtime checks for sorting and disjointness"),
+    strax.Option('merge_without_s1', default=True, infer_type=False,
+                 help="If true, S1s will be igored during the merging. "
+                      "It's now possible for a S1 to be inside a S2 post merging"),
+)
 class Peaks(strax.Plugin):
     """
     Merge peaklets and merged S2s such that we obtain our peaks
@@ -629,9 +859,9 @@ class Peaks(strax.Plugin):
     data_kind = 'peaks'
     provides = 'peaks'
     parallel = True
-    save_when = strax.SaveWhen.NEVER
+    save_when = strax.SaveWhen.EXPLICIT
 
-    __version__ = '0.1.1'
+    __version__ = '0.1.2'
 
     def infer_dtype(self):
         return self.deps['peaklets'].dtype_for('peaklets')
@@ -639,13 +869,24 @@ class Peaks(strax.Plugin):
     def compute(self, peaklets, merged_s2s):
         # Remove fake merged S2s from dirty hack, see above
         merged_s2s = merged_s2s[merged_s2s['type'] != FAKE_MERGED_S2_TYPE]
-
-        peaks = strax.replace_merged(peaklets, merged_s2s)
+        
+        if self.config['merge_without_s1']:
+            is_s1 = peaklets['type'] == 1
+            peaks = strax.replace_merged(peaklets[~is_s1], merged_s2s)
+            peaks = strax.sort_by_time(np.concatenate([peaklets[is_s1],
+                                                       peaks]))
+        else:
+            peaks = strax.replace_merged(peaklets, merged_s2s)
 
         if self.config['diagnose_sorting']:
             assert np.all(np.diff(peaks['time']) >= 0), "Peaks not sorted"
-            assert np.all(peaks['time'][1:]
-                          >= strax.endtime(peaks)[:-1]), "Peaks not disjoint"
+            if self.config['merge_without_s1']:
+                to_check = peaks['type'] != 1
+            else:
+                to_check = peaks['type'] != FAKE_MERGED_S2_TYPE
+
+            assert np.all(peaks['time'][to_check][1:]
+                            >= strax.endtime(peaks)[to_check][:-1]), "Peaks not disjoint"
         return peaks
 
 
@@ -666,33 +907,53 @@ class PeaksHighEnergy(Peaks):
 
 
 @numba.jit(nopython=True, nogil=True, cache=True)
-def get_tight_coin(hit_max_times, peak_max_times, left, right):
-    """Calculates the tight coincidence
+def get_tight_coin(hit_max_times, hit_channel, peak_max_times, left, right,
+                   channels=(0, 493)):
+    """Calculates the tight coincidence based on PMT channels.
 
     Defined by number of hits within a specified time range of the
     the peak's maximum amplitude.
     Imitates tight_coincidence variable in pax:
     github.com/XENON1T/pax/blob/master/pax/plugins/peak_processing/BasicProperties.py
+
+    :param hit_max_times: Time of the hit amplitude in ns.
+    :param hit_channel: PMT channels of the hits
+    :param peak_max_times: Time of the peaks maximum in ns.
+    :param left: Left boundary in which we search for the tight
+        coincidence in ns.
+    :param right: Right boundary in which we search for the tight
+        coincidence in ns.
+    :param channel_range: (min/max) channel for the corresponding detector.
+
+    :returns: n_coin_channel of length peaks containing the
+        tight coincidence.
     """
     left_hit_i = 0
-    n_coin = np.zeros(len(peak_max_times), dtype=np.int16)
+    n_coin_channel = np.zeros(len(peak_max_times), dtype=np.int16)
+    start_ch, end_ch = channels
+    channels_seen = np.zeros(end_ch-start_ch+1, dtype=np.bool_)
 
     # loop over peaks
     for p_i, p_t in enumerate(peak_max_times):
-
+        channels_seen[:] = 0
         # loop over hits starting from the last one we left at
         for left_hit_i in range(left_hit_i, len(hit_max_times)):
 
             # if the hit is in the window, its a tight coin
             d = hit_max_times[left_hit_i] - p_t
-            if (-left < d) & (d < right):
-                n_coin[p_i] += 1
+            if (-left <= d) & (d <= right):
+                channels_seen[hit_channel[left_hit_i]-start_ch] = 1
 
             # stop the loop when we know we're outside the range
             if d > right:
+                n_coin_channel[p_i] = np.sum(channels_seen)
                 break
+        
+        # Add channel information in case there are no hits beyond 
+        # the last peak:
+        n_coin_channel[p_i] = np.sum(channels_seen)
 
-    return n_coin
+    return n_coin_channel
 
 
 @numba.njit(cache=True, nogil=True)
